Refactor main.js to modularize services and handlers, add 'diff' package to dependencies, and improve IPC handler organization. Update package-lock.json and package.json accordingly.
