+++ src/config/constants.js
@@ -1,0 +1,36 @@
+// Language shortcuts mapping
+const LANGUAGE_SHORTCUTS = {
+  'e': 'English',
+  'p': 'Polish',
+  'c': 'Chinese',
+  'g': 'German',
+  'f': 'French',
+  's': 'Spanish',
+  'r': 'Russian',
+  'i': 'Italian',
+  'j': 'Japanese',
+  'h': 'Hindi',
+  'a': 'Arabic',
+  't': 'Portuguese',
+};
+
+// Default window dimensions
+const WINDOW_CONFIG = {
+  width: 800,
+  height: 800,
+  minWidth: 800,
+  minHeight: 600,
+};
+
+// Default hotkey
+const DEFAULT_HOTKEY = 'CommandOrControl+Shift+C';
+
+// Translation hotkey
+const TRANSLATION_HOTKEY = 'CommandOrControl+Shift+T';
+
+module.exports = {
+  LANGUAGE_SHORTCUTS,
+  WINDOW_CONFIG,
+  DEFAULT_HOTKEY,
+  TRANSLATION_HOTKEY,
+};
